Add logic for cart total and item display
Deleted cart items will disappear
When item is deleted, new total price is shown
All cart actions moved to their own file (cart.js)
