+++ assets/scripts/example.js
@@ -1,3 +1,9 @@
 'use strict';
 
-module.exports = true;
+const test = function() {
+  console.log('test');
+};
+
+module.exports = {
+  test
+};
