+++ assets/scripts/templates/cart.handlebars
@@ -1,5 +1,5 @@
 {{#each cart}}
-  <div class="well well-sm">
+  <div class="well well-sm line-item" data-id={{productid}}>
     <div class="row">
       <div class="col-xs-3">
         <img src={{image}} class="img-responsive img-cart"/>
