+++ assets/styles/index.scss
@@ -43,7 +43,7 @@ ul {
 
 .total-row {
   background-color: #cdfcc9;
-  padding: 5px 0;
+  padding: 5px;
 }
 
 .cart-total-line {
