+++ index.html
@@ -117,7 +117,7 @@
           <div class="cartDisplay">
 
           </div>
-          <div class="has-items total-row">
+          <div class="has-items total-row hidden">
             <h3 class="text-right has-items cart-total-line">Total: $<span id="cart-total"></span>  </h3>
           </div>
         </div>
